capdev: Have another thread to write control requests
Deadlock is observed when both processes are writing into pipes. Maybe
this is because of the small size of the buffer. However, it will be
safe to separate writing and reading thread to avoid the deadlock.
I don't want to have the additional thread on Linux and macOS, this
commit only affects for Windows.