Add a property async_compensation to switch asynchronous SRC
This commit implements a property to enable asynchronous sample rate
conversion (SRC) in the PR obsproject/obs-studio#6351.
This commit relies on the PR so that the implementation is disabled by
default.
Assuming the user of this plugin has enough knowledge for asynchronous
sample rate and genlock, the option is disabled by default.